% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build the website}
\usage{
build(dir = getwd(), ...)
}
\arguments{
\item{dir}{directory containing the website (contains "source")}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Build the website
}

