% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Remove the build directory}
\usage{
clean(dir = getwd())
}
\arguments{
\item{dir}{directory containing the website (contains "source" and "build")}
}
\description{
Remove the build directory
}
