% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{compare_state}
\alias{compare_state}
\title{Compare two directory states}
\usage{
compare_state(old, new)
}
\arguments{
\item{old}{previous state}

\item{new}{current state}
}
\value{
list containing number of changes \code{n}, and files which have been \code{added}, \code{deleted}, \code{modified} and overall \code{changed}
}
\description{
Compare two directory states
}
