% Generated by roxygen2 (4.0.0): do not edit by hand
\name{display_list}
\alias{display_list}
\title{Create an HTML list}
\usage{
display_list(x, type = c("ul", "ol"), ...)
}
\arguments{
\item{x}{an R object (a vector or a list), each element of which will be wrapped in \code{<li>} tags and inserted into a \code{<ul>}/\code{<ol>} tag. If x is a list of lists, it will result in nested \code{<ul>}/\code{<ol>} lists}

\item{type}{the type of HTML list : unordered or ordered}

\item{...}{passed to \code{\link{shiny::tags}}}
}
\description{
Create an HTML list
}
\examples{
display_list(c("a", "b", "c"))
display_list(c("a", "b", "c"), type="ol")
display_list(list("a", list("a.1", "a.2"), "b"))
display_list(list("a", list("<a href='foo.html'>a.1</a>", "a.2"), "b"))
}

