% Generated by roxygen2 (4.0.0): do not edit by hand
\name{display_table}
\alias{display_table}
\title{Create an HTML table}
\usage{
display_table(x, digits = NULL, ...)
}
\arguments{
\item{x}{a data.frame, 2D matrix, table, etc. to be displayed as an HTML table}

\item{digits}{number of digits to display for numeric columns; default is all digits}

\item{...}{passed to \code{\link{shiny::tags}$table}}
}
\description{
Create an HTML table
}
\examples{
d <- data.frame(a=1:3, b=runif(3))
display_table(d)
display_table(d, digits=3)
}

