% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_role.R
\name{file_role}
\alias{file_role}
\title{Guess the role of files from their path and extension}
\usage{
file_role(path)
}
\arguments{
\item{path}{path to \emph{one} file, relative to the root of the site.}
}
\value{
A string with the role of the file
}
\description{
Guess the role of files from their path and extension
}
\examples{
file_role("code.R")
file_role("code.r")
file_role("my_page.Rmd")
file_role("my_page.md")
file_role("my_page.html")
file_role("assets/style.css")
file_role("assets/logo.png")
file_role("templates/main.html")

}

