% Generated by roxygen2 (4.0.2): do not edit by hand
\name{form_bootstrap}
\alias{checkbox_bs}
\alias{form_bootstrap}
\alias{input_bs}
\alias{label_bs}
\alias{radio_bs}
\alias{select_bs}
\alias{submit_bs}
\alias{textarea_bs}
\title{Bootstrap form elements}
\usage{
form_bootstrap(action, ..., method = "post", `accept-charset` = "utf-8")

label_bs(name, label)

input_bs(name, label = capitalise(name), type = "text", help = NULL,
  required = FALSE, ...)

textarea_bs(name, label = capitalise(name), rows = 3, help = NULL,
  required = FALSE, ...)

radio_bs(name, choices, label = capitalise(name), ...)

checkbox_bs(name, choices, label = capitalise(name), ...)

select_bs(name, choices, label = capitalise(name), selected = NULL,
  multiple = FALSE, ...)

submit_bs(label = "Submit", class = "btn btn-default", ...)
}
\arguments{
\item{action}{action link for the form}

\item{...}{passed to the corresponding \code{\link[shiny]{tags}} function}

\item{method}{post or get}

\item{accept-charset}{Unicode by default}

\item{name}{name of the element, used to process form data in php}

\item{label}{label of the element, visible on the page; by default, a capitalised version of the name}

\item{type}{of input. Is checked against valid HTML5 input types}

\item{help}{help text, placed under the element}

\item{required}{boolean, wether the input is required or not}

\item{rows}{number of rows shown by default in multiline text areas}

\item{choices}{a vector of possible values; if named, the content is the value tag and the names is displayed}

\item{class}{class of the submit button}
}
\description{
Bootstrap form elements
}
\seealso{
\code{\link[shiny]{tags}}
}

