% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_build_path}
\alias{get_build_path}
\title{Change the path of a file from source to build}
\usage{
get_build_path(path)
}
\arguments{
\item{path}{path to the file to process, in the source directory}
}
\description{
Change the path of a file from source to build
}
\keyword{internal}

