% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_info}
\alias{get_info}
\title{Get modification date and size of a file}
\usage{
get_info(...)
}
\arguments{
\item{...}{passed to \code{\link[base]{file.info}}, i.e. character vectors containing file paths.}
}
\description{
Get modification date and size of a file
}

