% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list_site_files}
\alias{list_site_files}
\title{List relevant files of a site}
\usage{
list_site_files(dir = getwd(), exclude = glob2rx("*.xls|xlsx|csv"))
}
\arguments{
\item{dir}{directory containing the website (contains "source")}

\item{exclude}{character string (regular expression) designating files to exclude from being copied in the destination directory. By default, exclude xls, xlsx and csv files which are usually used to contain data from which the website is generated. Set it to NULL to exclude nothing}
}
\description{
List all files then exclude some file type unwanted in the output and reorder files to be processed in a specified order
}

