% Generated by roxygen2 (4.0.1): do not edit by hand
\name{obscure_email}
\alias{obscure_email}
\title{Obscure email address}
\usage{
obscure_email(email)
}
\arguments{
\item{email}{character string which is an email address}
}
\description{
Obscure email address by converting characters to their unicode hex code. This procedure will indeed fool many, if not most, address-harvesting bots, but it definitely won't fool all of them.
}

