% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ol}
\alias{ol}
\title{Create an HTML ordered list}
\usage{
ol(x, ...)
}
\arguments{
\item{x}{an R object (a vector or a list), each element of which will be wrapped in \code{<li>} tags and inserted into a \code{<ol>} tag. If x is a list of lists, it will result in nested \code{<ol>} lists}

\item{...}{passed to \code{\link{shiny::tags}$ol}}
}
\description{
Create an HTML ordered list
}
\examples{
ol(c("a", "b", "c"))
ol(list("a", list("a.1", "a.2"), "b"))
}

