% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\alias{process.code}
\alias{process.deleted}
\alias{process.other}
\alias{process.template}
\title{Process website files}
\usage{
process(path, ...)

\method{process}{deleted}(path, ...)

\method{process}{other}(path, ...)

\method{process}{code}(path, ...)

\method{process}{template}(path, ...)
}
\arguments{
\item{path}{path to the file to process, in the source directory}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Process a file from the source directory into the build directory according to its role, as determined by \code{\link{file_role}}
}

