% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read site configuration}
\usage{
read_config(dir = getwd())
}
\arguments{
\item{dir}{to site directory, by default the current directory}
}
\value{
A list, in which various options are defined
}
\description{
Source the \code{config.R} file at the root of the site, when it exists, which contains options.
}

