% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render}
\alias{render}
\title{Render a template file inside a layout}
\usage{
render(file, layout, dest, ...)
}
\arguments{
\item{file}{name of a file containing content to render through \code{\link{render_content}} and the insert into a layout template. The extension of the file determines which rendering function is used. Allowed extensions are
\describe{
\item{".brew"}{brew template, rendered by \code{\link[brew]{brew}}}
\item{".rmd"}{R markdown template, rendered by \code{\link[knitr]{knit2html}}}
\item{".rhtml"}{R html template, rendered by \code{\link[knitr]{knit}}}
\item{".md" or ".markdown"}{markdown template, rendered by \code{\link[markdown]{markdownToHTML}}}
\item{".html" or ".php"}{HTML or PHP code, passed through to be inserted in the layout template}
}
NB: the case of the extension does not matter, e.g. \code{"Rmd"} and \code{"rmd"} are equivalent}

\item{layout}{path to layout template file}

\item{dest}{path to destination directory}

\item{...}{passed to \code{\link{render_content}}}
}
\description{
Render a template file inside a layout
}

