% Generated by roxygen2 (4.0.0): do not edit by hand
\name{render_content}
\alias{render_content}
\title{Render markup from a templating engine into HTML}
\usage{
render_content(text, ...)
}
\arguments{
\item{text}{character string with markup to feed to the appropriate conversion engine. Allowed markup are:
\itemize{
\item \code{\link{brew}}, with extension \code{.brew}
\item \code{R markdown}, with extensions \code{.Rmd} or \code{.rmd}
\item \code{markdown}, with extensions \code{.md} or \code{.markdown}
\item \code{R HTML} for \code{\link{knitr}}, with extensions \code{.Rhtml} or \code{.rhtml}
}}
}
\description{
Render markup from a templating engine into HTML
}

