% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scandir}
\alias{scandir}
\title{List files and directories}
\usage{
scandir(dir, pattern = NULL, type = c("ul", "ol"), ...)
}
\arguments{
\item{dir}{path to the directory of interest}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned (as in \code{\link{list.files}} on which scandir is based)}

\item{type}{of list to produce: unordered (\code{ul}) or ordered(\code{ol})}

\item{...}{passed to \code{\link{list.files}}}
}
\description{
Displays the content of the provided directory as an HTML list of links. Can display only certain files (e.g. those with a specific extension or prefix) using pattern matching. This function is somewhat equivalent to the PHP and Shell functions of the same name.
}

