% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve}
\alias{serve}
\title{Serve the local website}
\usage{
serve(dir = getwd())
}
\arguments{
\item{dir}{directory containing the website (contains "source" and "build")}
}
\description{
Serve the local website
}
