% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{toggle_attribute}
\alias{toggle_attribute}
\title{Toggle a logical value into an attribute for an HTML tag}
\usage{
toggle_attribute(x)
}
\arguments{
\item{x}{logical value}
}
\description{
Attributes are set in HTML tags by setting them to NA. This is not very human-readable. It is more understandable to set them to TRUE to set the attribute and to FALSE not to set it.
}
\examples{
library("shiny")
tags$input(name="Foo", required=toggle_attribute(FALSE))
tags$input(name="Foo", required=NULL)
tags$input(name="Foo", required=toggle_attribute(TRUE))
tags$input(name="Foo", required=NA)
}
