% Generated by roxygen2 (4.0.0): do not edit by hand
\name{toHTML}
\alias{toHTML}
\title{Convert markup from a templating engine into HTML}
\usage{
toHTML(text, ...)
}
\arguments{
\item{text}{character string with markup to feed to the appropriate conversion engine. Allowed markup are:
- \code{\link{brew}}, with extension \code{.brew}
- \code{R markdown}, with extensions \code{.Rmd} or \code{.rmd}
- \code{R HTML} for \code{\link{knitr}}, with extensions \code{.Rhtml} or \code{.rhtml}}
}
\description{
Convert markup from a templating engine into HTML
}

