% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ul}
\alias{ul}
\title{Create an HTML unordered list}
\usage{
ul(x, ...)
}
\arguments{
\item{x}{an R object (a vector or a list), each element of which will be wrapped in \code{<li>} tags and inserted into a \code{<ul>} tag. If x is a list of lists, it will result in nested \code{<ul>} lists}

\item{...}{passed to \code{\link{shiny::tags}$ul}}
}
\description{
Create an HTML unordered list
}
\examples{
ul(c("a", "b", "c"))
ul(list("a", list("a.1", "a.2"), "b"))
}

