% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build}
\alias{build}
\title{Build the website}
\usage{
build(dir = getwd(), ...)
}
\arguments{
\item{dir}{directory containing the website (contains "source")}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Build the website
}

