% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean}
\alias{clean}
\title{Remove the build directory}
\usage{
clean(dir = getwd(), ...)
}
\arguments{
\item{dir}{directory containing the website (contains "source" and "build")}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Remove the build directory
}

