% Generated by roxygen2 (4.0.0): do not edit by hand
\name{display_table}
\alias{display_table}
\title{Display an HTML table}
\usage{
display_table(x, ...)
}
\arguments{
\item{x}{a data.frame, 2D matrix, table, etc. to be displayed as an HTML table}

\item{...}{passed to \code{\link{shiny::tags}$table}}
}
\description{
Display an HTML table
}

