% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init}
\alias{init}
\title{Initialise an yssr website}
\usage{
init(dir = getwd(), ...)
}
\arguments{
\item{dir}{name/path to a directory to hold the website. Will be created if it does not exists}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Initialise an yssr website
}

