% Generated by roxygen2 (4.0.0): do not edit by hand
\name{init}
\alias{init}
\title{Initialise an yssr website}
\usage{
init(dir = getwd(), ...)
}
\arguments{
\item{dir}{name/path to a directory to hold the website. Will be created if it does not exists}
}
\description{
Initialise an yssr website
}

