% Generated by roxygen2 (4.0.2): do not edit by hand
\name{link}
\alias{link}
\title{Create HTML links}
\usage{
link(url, text = url, ...)
}
\arguments{
\item{url}{vector of character strings containing the url to which the links point}

\item{text}{vector of character strings containing the text of the links; what is displayed}

\item{...}{passed to \code{link[shiny]{a}}}
}
\description{
Create \code{a} tag(s). The function is vectorised over the inputs.
}

