% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mailto}
\alias{mailto}
\title{Create mailto links}
\usage{
mailto(address, text = NULL, obscure = TRUE, ...)
}
\arguments{
\item{address}{vector of character strings containing email addresses}

\item{text}{vector of character strings containing text of for the links. By default, the email address}

\item{obscure}{wether to obscure the email address with \code{\link{obscure_email}}}

\item{...}{passed to \code{link[shiny]{a}}}
}
\description{
Create \code{a} tag(s) with \code{href=mailto}. The function is vectorised over the inputs.
}
\seealso{
\code{\link{obscure_email}}
}

