% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_content}
\alias{render_content}
\alias{render_content.brew}
\alias{render_content.rmd}
\alias{render_content.md}
\alias{render_content.markdown}
\alias{render_content.rhtml}
\alias{render_content.default}
\title{Render markup from a templating engine into HTML}
\usage{
render_content(text, ...)

\method{render_content}{brew}(text, ...)

\method{render_content}{rmd}(text, options = "+smartypants", ...)

\method{render_content}{md}(text, options = "+smartypants", ...)

\method{render_content}{markdown}(text, options = "+smartypants", ...)

\method{render_content}{rhtml}(text, ...)

\method{render_content}{default}(text, ...)
}
\arguments{
\item{text}{character string with markup to feed to the appropriate conversion engine.}

\item{...}{further arguments passed to or from other methods.}

\item{options}{options passed to \code{\link[markdown]{markdownToHTML}}}
}
\description{
Render markup from a templating engine into HTML
}
