% Generated by roxygen2 (4.0.2): do not edit by hand
\name{watch}
\alias{watch}
\title{Watch the website source for changes and process files accordingly}
\usage{
watch(dir = getwd(), interval = 1, ...)
}
\arguments{
\item{dir}{directory containing the website (contains "source")}

\item{interval}{interval, in s, at which to watch for changes}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Watch the website source for changes and process files accordingly
}

